+++ plugins/transport-grpc/src/main/java/org/opensearch/plugin/transport/grpc/proto/response/BulkItemResponseProtoUtils.java
@@ -11,9 +11,6 @@ import org.opensearch.action.DocWriteResponse;
 import org.opensearch.action.bulk.BulkItemResponse;
 import org.opensearch.action.update.UpdateResponse;
 import org.opensearch.index.get.GetResult;
-import org.opensearch.plugin.transport.grpc.proto.DocWriteResponseProtoUtils;
-import org.opensearch.plugin.transport.grpc.proto.ExceptionProtoUtils;
-import org.opensearch.plugin.transport.grpc.proto.GetResultProtoUtils;
 import org.opensearch.protobuf.ErrorCause;
 import org.opensearch.protobuf.Item;
 import org.opensearch.protobuf.NullValue;
